+++ R/byosrmRoute.R
@@ -28,8 +28,7 @@
 
 
 byosrmRoute <- function(src_dst, crs){
-        assertthat::assert_that(.x = curl::has_internet() , 
-                                msg = "No internet access was detected. Please check your connection.")
+        EPP::osrm_ok()
         getOption("osrm.server") == "https://routing.openstreetmap.de/"
         pop_s <- SpatialPoints(src_dst[ ,1:2], proj4string = crs)## transform pop to spatial object
         centers_s <- SpatialPoints(src_dst[ ,3:4], proj4string = crs)## transform centers to spatial object